CORE-731 Refactor App Launch info and param steps into separate files.
