+++ src/components/apps/launch/constants.js
@@ -1,0 +1,3 @@
+export default {
+    PARAM_TYPE: { DOUBLE: "Double", INFO: "Info", INTEGER: "Integer" },
+};
